Move high-level interface out to highlevel.jl, prop manip to properties.jl
